refactor(musicbrainz): move MB limiter singleton next to its only caller
`musicbrainz_limiter()` and its rate constant don't belong in
`catalog/common/rate_limit.py` -- that module is the generic
RedisRateLimiter primitive shared by anyone who needs a Redis-backed
throttle. The MB-specific singleton lives in `catalog/sites/musicbrainz.py`
alongside the MusicBrainzDownloader that consumes it.

While moving it, lower the rate to MusicBrainz' documented public-API
guideline of 1 req/s/IP. The 50 req/s/IP ceiling we were running near is
only available to negotiated high-volume clients; assuming it would
starve other well-behaved tools sharing our egress IP.
